Complete the move of all OpenSSL functions to primitives file
