+++ include/common.h
@@ -54,6 +54,29 @@ operator!=(const T& lhs, const T& rhs) {
 }
 
 ///
+/// CipherSuite and Signature identifiers
+///
+
+enum struct CipherSuite : uint16_t
+{
+  P256_SHA256_AES128GCM = 0x0000,
+  P521_SHA512_AES256GCM = 0x0010,
+  X25519_SHA256_AES128GCM = 0x0001,
+  X448_SHA512_AES256GCM = 0x0011
+};
+
+size_t suite_nonce_size(CipherSuite suite);
+size_t suite_key_size(CipherSuite suite);
+
+enum struct SignatureScheme : uint16_t
+{
+  P256_SHA256 = 0x0403,
+  P521_SHA512 = 0x0603,
+  Ed25519 = 0x0807,
+  Ed448 = 0x0808
+};
+
+///
 /// Error types
 ///
