+++ src/state.cpp
@@ -1,5 +1,4 @@
 #include "state.h"
-#include "primitives.h"
 
 #define DUMMY_SIG_SCHEME SignatureScheme::P256_SHA256
 
@@ -75,11 +74,11 @@ State::State(const std::vector<ClientInitKey>& my_client_init_keys,
 
   // Decrypt the GroupInfo
   auto first_epoch = FirstEpoch::create(_suite, key_pkg.init_secret);
-  auto group_info_data = primitive::open(_suite,
-                                         first_epoch.group_info_key,
-                                         first_epoch.group_info_nonce,
-                                         {},
-                                         welcome.encrypted_group_info);
+  auto group_info_data = open(_suite,
+                              first_epoch.group_info_key,
+                              first_epoch.group_info_nonce,
+                              {},
+                              welcome.encrypted_group_info);
   auto group_info = tls::get<GroupInfo>(group_info_data, _suite);
 
   // Verify the singature on the GroupInfo
@@ -635,11 +634,11 @@ State::encrypt(const MLSPlaintext& pt)
   auto sender_data_aad_val = sender_data_aad(
     _group_id, _epoch, pt.content.inner_type(), sender_data_nonce);
 
-  auto encrypted_sender_data = primitive::seal(_suite,
-                                               _keys.sender_data_key,
-                                               sender_data_nonce,
-                                               sender_data_aad_val,
-                                               sender_data.bytes());
+  auto encrypted_sender_data = seal(_suite,
+                                    _keys.sender_data_key,
+                                    sender_data_nonce,
+                                    sender_data_aad_val,
+                                    sender_data.bytes());
 
   // Compute the plaintext input and AAD
   // XXX(rlb@ipv.sx): Apply padding?
@@ -652,7 +651,7 @@ State::encrypt(const MLSPlaintext& pt)
                          encrypted_sender_data);
 
   // Encrypt the plaintext
-  auto ciphertext = primitive::seal(_suite, keys.key, keys.nonce, aad, content);
+  auto ciphertext = seal(_suite, keys.key, keys.nonce, aad, content);
 
   // Assemble the MLSCiphertext
   MLSCiphertext ct;
@@ -681,11 +680,11 @@ State::decrypt(const MLSCiphertext& ct)
   // Decrypt and parse the sender data
   auto sender_data_aad_val = sender_data_aad(
     ct.group_id, ct.epoch, ct.content_type, ct.sender_data_nonce);
-  auto sender_data = primitive::open(_suite,
-                                     _keys.sender_data_key,
-                                     ct.sender_data_nonce,
-                                     sender_data_aad_val,
-                                     ct.encrypted_sender_data);
+  auto sender_data = open(_suite,
+                          _keys.sender_data_key,
+                          ct.sender_data_nonce,
+                          sender_data_aad_val,
+                          ct.encrypted_sender_data);
 
   tls::istream r(sender_data);
   LeafIndex sender;
@@ -722,8 +721,7 @@ State::decrypt(const MLSCiphertext& ct)
                          ct.authenticated_data,
                          ct.sender_data_nonce,
                          ct.encrypted_sender_data);
-  auto content =
-    primitive::open(_suite, keys.key, keys.nonce, aad, ct.ciphertext);
+  auto content = open(_suite, keys.key, keys.nonce, aad, ct.ciphertext);
 
   // Set up a new plaintext based on the content
   return MLSPlaintext{ _suite, _group_id,       _epoch,
